Allow other strings as cursors
In #89 itwas pointed out that using other scalars as cursors would be useful,
since clients might want to provide a custom scalar the serializes as a String.
Update the spec to ensure that is possible.

This change keeps the requirement that cursors serialize as strings, which
be strongenough to allow clients to handle different servers. It no longer
mandates a type, so servers can just use String if they prefer, or define
a custom scalar if they prefer.

Fixes #89.