printRelayQuery returns text plus variables
Summary: For OSS the printer needs to be able to synthesize variables when an argument has a complex type. An example of the output is as follows:

  query UnknownFile($0:CheckinSearchInput) {
    checkinSearchQuery(query:$0) {
      ...
    }
  }

Inlining a value of type `CheckinSearchInput` requires understanding the shape of this input object, and knowledge of any (nested) keys whose type is `Enum` (which are printed differently than strings). The options for sending such queries to the server are a) store metadata about the shape of `InputObject` types and which keys are enums or b) use variables and send the values as plain JSON (and let the server handle decoding types appropriately).

This diff implements option b. The OSS version of `printRelayQuery` (implemented [[ https://github.com/josephsavona/relay/commit/e2cb461f6222df8fbe95a0ffcfb90636c5b89a07 | here ]]) synthesizes variables for any non-scalar argument types.

Reviewed By: @yungsters

Differential Revision: D2368818