Use HTTPS URL for starter kit
This is lower friction than the SSH URL because it works even without a
GitHub account and a configured SSH key.

Note there is one other SSH URL in the repo that I did not change in
`website/README.md`. That one specifically deals with publishing the
website, for which you need push access, and SSH is going to be more
convenient for that.