better support for optmistic edges
Summary: Optimistic mutations were not working for the TODO example in oss.
Generally we read the prepended edges with `first` calls and read the appended edges with the edges for `last` calls. In the TODO example, we read using the `first` call, and added new ones at the bottom of the list. The new todos were not showing up because we never check the appended edges when reading with the `first` call. We should include the appended edges if there is no more next page and we should include the prepended edges if there is no more previous page.
This will apply appended when there are no more next page and prepended edges when there areno more previous page

Reviewed By: @josephsavona

Differential Revision: D2407213