Relay: Unify Forks
Summary: Part of the pain we're currently experiencing with managing two repositories (Facebook and GitHub) is the magic `__forks__` directory that allows us to use one version of a module internally and another version of the module externally.

This diff moves the two versions of every forked module out of `__forks__` and renames them to either `*RelayFB*` or `*RelayOSS*`.

The `__forks__` directory still exists, but the files there should almost never need to be changed. Internal versions of the fork simply point to the `*RelayFB*` module, while the open source versions simply point to the `*RelayOSS*` module. Also, the `printRelay*Query-test.js` files are still in `__forks__` because they inherent test what makes us need the fork to begin with (i.e. different versions of GraphQL).

Reviewed By: @josephsavona

Differential Revision: D2409517