Renaming RelayRecordStatus to RelayRecordState
Summary: I want to reuse the name `RelayRecordStatus` for whether a node is part of a optimistic mutation or if an mutation on the node has resulted in an error. These are currently stored under `__status__` and `__hasError__`. I plan to consolidate some of the logic in RelayRecordStatus. Renaming the current RecordStatus to State since node has to be in one and only one of these states (I'm thinking state machine). Want to resuse `RelayRecordStatus` since a record is capable of multiple statuses.

Reviewed By: @josephsavona

Differential Revision: D2412702