Dont flatten mutation queries
Summary: `RelayMutationQuery` is not maintaining fragments from the original tracked queries - the call to `refragmentRelayQuery()` is too early and queries are still flattened afterward. This is now fixed and `flattenRelayQuery()` is only used where fields are manually added in a way that could cause duplicates (and this is now before refragmenting).

Addresses #234
Closes https://github.com/facebook/relay/pull/249

Reviewed By: @yungsters

Differential Revision: D2419716