babel-relay-plugin: disallow before/first after/last; inline actual errors
Summary: Makes the use of `field(before: .., first: ..)` or after/last an error. This combination is only supported when generated by Relay during the diffing process.

Also, the previous version replaced invalid GraphQL templates with a function that throws - while I was at it, I changed this to include the actual validation error message.

Addresses #247
Closes https://github.com/facebook/relay/pull/252

Reviewed By: @yuzhi

Differential Revision: D2419715