buildRQL ignores unknown route params
Summary: D2391474 added support for "streamlined" route queries definitions of the form ##() => Relay.QL`query { viewer }`##. The implementation pre-allocated fragment references to the root fragment in case they were needed (in case of streamlined query definitions), and passed in *all* route params as variables to the root fragment. This broke the invariant in `RelayFragmentReference` that all variables must be undefined, because route params can be optional.

The fix here is to a) create fragment references only when the route query actually uses the new style and b) only pass route params that correspond to a variable in the root component.

Reviewed By: @yungsters

Differential Revision: D2415143