Inject source code into the prototyping tools via the hash instead
Summary: So, it turns out that GitHub's varnish cluster is configured with a maximum URL length. This thwarted the deployment of our homepage examples, since we inject the source code into them via the URL.

This PR makes the injection happen via the hash instead of `location.search`, where the server can't touch it.

![giphy](https://cloud.githubusercontent.com/assets/13243/9769176/67f62d68-56dd-11e5-9498-8213d01bea9c.gif)

Closes https://github.com/facebook/relay/pull/270

Reviewed By: @yungsters

Differential Revision: D2425352