Removing warning
Summary: Joe recently added warning in the transform to show up when people write
queries with first.before and last.after so we dont need this warning anymore.
We still have to keep the if block since adding logic to read those queries is not trivial and it's only used
to try to reuse edges when we write data from diff queries. We don't about to worry about these types of queries
on the read path since we don't allow these queries in the components.

Reviewed By: @yungsters

Differential Revision: D2425425