Use ReactDOM for batchedUpdates
Summary: Batching is unique to each renderer so we can't access it on the React module and have to use ReactDOM or ReactNative as appropriate. This change means Relay OSS won't work with React Native, but that's already true due to fbjs and flow problems. Internally for RN, we'll shim ReactDOM as just `{unstable_batchedUpdates: ReactUpdates.batchedUpdates}`. so that this continues to work. We'll see if we can make batching automatic in the next version of React.

Reviewed By: @sebmarkbage

Differential Revision: D2428361