Deploy Flow v0.15.0 to www
Summary: This diff not only updates the `.flowconfig` but also fixes a lot of
Flow errors in www. Some of these fixes must go out with 0.15.0. Things like

* Removing unused error suppressions due to Flow issues being fixed, errors
  moved, Flow being smarter, etc
* Modifying some code to use generators

There are some misc fixes that could work with 0.14.0 but I was have a rough
time separating them out so I just squished them together.

Reviewed By: @jeffmo

Differential Revision: D2429385