Relay: Create Root Records w/ Queued Store
Summary: When we're constructing root records in `RelayRootContainer`, we require a reference to the store because we need to use the root call mapping (e.g. from `me` to `node(id:123)`).

However, `RelayRootContainer` is currently using `getRecordStore()` instead of `getQueuedStore()`. This means that we're not using the correct root call mapping (e.g. missing out on cached or optimistic root call mappings).

Reviewed By: @Invalid auth token., @josephsavona

Differential Revision: D2433873