diffRelayQuery creates fragmented queries
Summary: Addresses the issue in #305: `diffRelayQuery` creates top-level queries for missing information but did not wrap the fields in fragments based on type. Rather than use `refragmentRelayQuery` which requires a full traversal of the query, this change relies on the fact that input queries are already correctly fragmented, and only needs to group the top-level fields into fragments by type.

The schema changes are required because otherwise `... on FeedUnit` would not be allowed in a `Node` context.
Closes https://github.com/facebook/relay/pull/316

Reviewed By: @wincent

Differential Revision: D2442000