fix module dependency on side effects from when it was required
Summary: This module depends on the initial value of `console.error` when it gets required and with inline require
it doesn't work as expected.
The fix is to check for the `isSpy` property instead, but it's not available on the edge version of jasmine, so I added that to the test as a way to warn when we update jasmine on jest.
Another alternative would be to use `jasmine.isSpy`, that works on future versions of jasmine, but I'm not sure how clean is that compared to the current implementation. Let me know your opinion.

Reviewed By: @spicyj

Differential Revision: D2445713