You can now send someone a permalink to a Relay playground
Summary: This will enable us to host an instance of the Relay playground on facebook.github.io whose permalinks can be shared. Being able to share playgrounds should make it easy for people to share ideas and bug reports.

This diff enables this while protecting against code injection vulnerabilities. Auto-execution is whitelisted for localhost and wherever a playground is being run in an `<iframe>` on the Relay website (we need this for the tutorial materials). Everywhere else, a mismatch between what you have cached locally and what's in the URL will disable auto-execution and trigger a confirmation dialog. The user will have an opportunity to evaluate the code before choosing to execute it.

![relay](https://cloud.githubusercontent.com/assets/13243/9966456/e9ff9324-5df1-11e5-8810-c5bc2bdd1db1.gif)

Closes https://github.com/facebook/relay/pull/342

Reviewed By: @yungsters

Differential Revision: D2457894