create transformRelayQueryPayload()
Summary: A lot of our tests have to write data into the store before testing some other logic. However, the payloads must be manually constructed using `generateRQLFieldAlias()`. It's not a big deal, but it's annoying.

So this PR adds a new module `transformRelayQueryPayload(query: RelayQuery.Node, payload: Payload): Payload` - it accepts payloads where the keys correspond to field names/aliases (which cannot be written into the store as-is), and returns an equivalent payload where keys are renamed to the serialization key of the corresponding field (which *can* be written to the store). See the unit test for this module for an example, and how it (slightly) simplifies tests that write to the store.

Closes https://github.com/facebook/relay/pull/341

Reviewed By: @yungsters

Differential Revision: D2457487