Move adding pageInfo for edges up one level
Summary: The current logic for adding `pageInfo` only works when it is able to directly check the parent *field* and verify that it is a connection and contains an `edges` field. @​taion fixed this by moving this logic to where we traverse any parent node and testing that the type is a connection.

Potential fix for #381.
Closes https://github.com/facebook/relay/pull/385

Reviewed By: @yungsters

Differential Revision: D2478089