Fix the result of hasNextPage and hasPreviousPage in specification.
Summary: The result from the hasPreviousPage and hasNextPage algorithms are backwards in the connection specification.  It was saying to return `false` if they are *more* than the requested number of elements, but in that case there is another page, so it hasn't reached the end of the set of edges.
Closes https://github.com/facebook/relay/pull/397

Reviewed By: @josephsavona

Differential Revision: D2485552