Fixed example projects building Relay twice on npm install
Summary: Fixes #367 where `npm install` on the projects in `example/` would build twice. This is due to the nature of `npm install` which recursively executes `npm install --production` on parent directories, including `relay/`.

Since the `preinstall` script of the projects already executes `npm install` on `relay/`, `npm install` runs twice on `relay/` therefore running the `build` script of `relay/` twice.

Adding `--ignore-scripts` to the `preinstall` script of the projects fixes this by not executing any scripts in the first `npm install` on `relay/`.
Closes https://github.com/facebook/relay/pull/382

Reviewed By: @kassens

Differential Revision: D2478364