Harmonize Star Wars example with relay-starter-kit
This and the other examples are intended to be minimal demos over a
minimal substrate of infrastructure. "relay-starter-kit" is basically
the template for that substrate.

In order to make it easier to bring across changes from the starter kit
to here, this commit makes the example app more like the starter kit, so
that if we want to update `server.js` in the future, for example, we can
just copy it over directly with no modifications.

Changes:

- Made schema filenames generic (`starWarsSchema` -> `schema` and
  `starWarsDatabase` -> `database`).
- Got rid of eslint, which the starter kit does not use and isn't really
  needed in such a tiny example app.

Test plan: `npm run update-schema`, `npm run start`, and hit site with
web browser.