Synchronize examples with starter kit, using explicit imports
Summary: Brings across:

- https://github.com/relayjs/relay-starter-kit/pull/36
- https://github.com/wincent/relay-starter-kit/pull/1

from the starter kit. Benefit here is explicitness and robustness, but the cost is that the webpack bundle may take longer to build (I'm going to look at tweaking the config though to not reparse React/Relay when bundling, if that can be done without [causing problems](http://stackoverflow.com/questions/32708927/how-do-you-move-from-stock-relay-to-relay-with-a-router)).

Closes https://github.com/facebook/relay/pull/404

Reviewed By: @kassens

Differential Revision: D2492602