Trigger rerender when optimistic mutation has been confirmed
Summary: Currently hasOptimisticUpdate reads from global state and can only be called when we rerender.
This means that render() is not pure which in turn causes a bug since RelayContainer.shouldComponentUpdate
wrongly returns false when something that was optimistic is no longer optimistic but all the fields are still the same.

This diff will write optimistic and error flags into `__status__` for the nodes. When the bitmaps change, it will
trigger render to be called.

Reviewed By: @josephsavona

Differential Revision: D2491659