Move subscription count handling to `RelayStoreData`.
Currently `GraphQLStoreQueryResolver` is aware of the
`RelayStoreGarbageCollection` and updates the subscription-count directly. This
unnecessarily make the `GraphQLStoreQueryResolver` aware of garbage collection.
Using `RelayStoreData` as a layer between decouples the components. Furthermore
this will help implementing `RelayStoreData#reset` since we can in the future
keep track of active subscriptions in `RelayStoreData` itself.