Relay: Test OSS Babel Plugin Internally
Summary: Makes the necessary changes for the open source `babel-relay-plugin` unit tests to be run internally and continuously.

Internally, we use a heuristic to minimize the overhead of transforming "third party" code. The added `@​fullSyntaxTransform` annotations allow us to bypass the heuristic (so we can continue to use ES201X).

Reviewed By: @kassens

Differential Revision: D2507216

fb-gh-sync-id: dd288cb8abcbca2549df0f5101037f24042914be