Relay: Fix Babel Plugin Test Schema
Summary: When I tried to add new fields to the `babel-relay-plugin` test schema, I got pretty confused because the schema does not really make sense. (When we open sourced it, we tried to bend the schema to the tests' wills because it was quickest.)

This revises the fixtures and schema to make a bit more sense (which meant using a lot more inline fragments). Also, sprinkled in all those trailing commas and proper GraphQL spacing.

Reviewed By: @kassens

Differential Revision: D2507233

fb-gh-sync-id: 380977c87aebcda4dd3fb0566848ea7dada78252