Relay: Use Typed AST in Babel Plugin
Summary: Overhauls the `babel-relay-plugin` by pulling out all logic related to traversing the GraphQL AST and cross-referencing it with the GraphQL schema into a new `RelayAST` module. The goals of this are:

 - Share printer code between GraphQL FB and OSS plugins.
 - Improve readability and maintainability of `RelayQLPrinter`.
 - Upgrade abstractions using GraphQL OSS terminology and concepts.

For the most part, this should be functionally equivalent. The main difference is that directives are now validated.

Reviewed By: @josephsavona

Differential Revision: D2507451

fb-gh-sync-id: 90ddb5a28d98e875c6a65e1713529a516084812d