Move record store type defs into Relay(Internal)Types
Summary: As part of changing the interface to the disk cache manager I realized that this type was buried in `RelayRecordStore`. Flow-typing `RelayDiskCacheManager` correctly requires importing that type, though, so this moves it to `RelayTypes`, and some other types sich as `Record` out of `RelayRecordStore` and into `RelayInternalTypes` to be reusable.

This is a prereq to D2514943

Reviewed By: @yungsters

Differential Revision: D2514758

fb-gh-sync-id: 3a2c461890d87a353c9f974b4210670d89523c48