Relay: Preserve Whitespace as Separators in Plugin
Summary: We're doing two strange things in the Babel plugin:

 - We `trim()` every template element.
 - By accident, we're appending a comma after every substitution.

We need to `trim()` because there cannot be whitespace before the `OperationType` (e.g. `query` or `mutation`). However, the rest of the whitespace can be preserved. In fact, they //should// be preserved because whitespace has meaning in GraphQL (as a separator).

Furthermore, we do not need to append a comma to every substitution because commas are insignificant.

Reviewed By: @josephsavona

Differential Revision: D2520901

fb-gh-sync-id: 079b55937ca2c0a189877baee0de08e88a0c6ff0