CacheManager: separate objects for caching query/mutation payloads
Summary: The goal is for Relay to send to field-level updates to the CLC when processing the results of a mutation. The simplest integration point I could find was to piggyback on the cache manager. This diff changes the cache manager API so that it returns different "transaction" objects for caching the results of queries and mutations. The idea is that in general these may be the same object (and write data directly to disk), but the CLC integration can return a custom mutation transaction that both writes to disk cache //and// updates the CLC.

Changes:
* Change CacheManager API
* Add `RelayDiskCacheManager.clear()` to force-clear the on-disk cache - this is mostly for testing purposes but I can imagine product use-cases too (note: this does not affect Relay's in-memory `cachedRecords`)
* Adds a bunch of helpers in `DliteWeb` for testing disk caching on the web.

The API for cache manager becomes:

```
type CacheManager = {
  getMutationWriter: () => CacheWriter;
  getQueryWriter: () => CacheWriter;
  readAllData: (...) => void;
};
type CacheWriter = {
  write*: (...) => void;
};
```

Two other options were:
* Pass a wrapped RelayRecordStore instance that delegated to the original one, but also logged calls out to the CLC. This seemed like a more complicated version of the approach above.
* Change RelayQueryWriter to return an array of change descriptions, which would be applied to the store in a second pass. This was a big change and could affect perf, so I was hesitant to do this without further testing and/or some other goal. Definitely something to consider for the future.

Reviewed By: @yungsters

Differential Revision: D2514943

fb-gh-sync-id: d57d3733bec255efa82bf4f73a93f8de372a1b95