Incremental Read Visitor
Summary: This add a new visitor. This visitor will traverse the query node using the store and return a list of leaf query nodes that we were not able to traverse due not having read the relavant nodes from disk cache. It will also short circurt and return `result.missingData: true` if we have determined that the query is not satisfiable even with disk data.
The resulting pendingNodes from this visitor will be queued up to be read from disk cache and recursively traversed when the nodes are ready.

Welcome naming suggestions!
Also I don't really like passing in the raw cachedRecords, but I need to tell `undefined` (we tried to read from cache, but there was nothing with in TTL) `!object.hasOwnProperty` (we never read it from disk cache yet, could have data we don't know about yet). Suggestions welcomed!

Reviewed By: @josephsavona

Differential Revision: D2519996

fb-gh-sync-id: 357818d17e2874aaf0d6eac76b45416a05ba1bc6