make babel-relay-plugin a peer dependency
Summary: Currently there is nothing to guarantee that the version of `babel-relay-plugin` being used by an application matches the version of Relay. This can cause features to break when an upgraded Relay expects metadata that is not provided by older versions of the plugin - see #470 for an example.

Making the plugin a peer dependency helps to ensure that versions match.
Closes https://github.com/facebook/relay/pull/471

Reviewed By: @yungsters

Differential Revision: D2546738

fb-gh-sync-id: 013ed09f98b65a3a52e3506caecf5d4a7f15ba0b