Relay: Create <RelayRenderer>
Summary: Creates a new `<RelayRenderer>` component that serves the same purpose as `<RelayRootContainer>` but has a more flexible API.

Instead of `<RelayRootContainer>`'s `render{Failure,Fetched,Loading}` props, the new `<RelayRenderer>` component takes a single `render` callback. This makes it easier to define the behavior in which data is ready but an error occurs while fetching the remaining data. The existing `<RelayRootContainer>` can be implemented on top of `<RelayRenderer>`.

Since the tests for `<RelayRootContainer>` used shallow renderer, I essentially had to rewrite all of the tests. In the course of rewriting the tests, I discovered a couple edge cases and fixed them:

- Previously, if `<RelayRootContainer>` was rendered with either a new `Component` or `route` before a request is sent, `renderLoading` would not be called for the new props. Now in `<RelayRenderer>`, this is fixed.
  - **Internal:** Previously when `Component` or `route` changed, `activeComponent` and `activeRoute` would be nulled out. It is now always the currently "active" component and route.
  - **Internal:** Previously when `Component` and `route` were null, `<StaticContainer>`'s `shouldUpdate` was false. It is now true.
- Previously, updating a `<RelayRootContainer>` with `forceFetch={true}` with the same `Component` and `route` as a pending request would not trigger the correct "force fetch" query. It will now correctly abort the previous request and send a new one.
  - **Subtle:** If the previous query was also a "force fetch", it will not abort it.
  - **Subtle:** If the previous query was a "force fetch" but `<RelayRenderer>` is updated with `forceFetch={false}`, it will not abort the previous request.

Reviewed By: josephsavona

Differential Revision: D2530826

fb-gh-sync-id: 7c39d568e2da308351ead23dcd5c0d3a806519aa