Replace babel-relay-plugin build script with node script
We've had reports of `npm install` not working on Windows because our
`build-lib` script makes platform-specific assumptions:

https://github.com/facebook/relay/issues/484

This commit rewrites it in JS, which in theory should run everywhere
`node` runs. (Note, I haven't actually tested this on Windows yet,
although I have confirmed that it continues to do the right thing on OS
X.)

Thing is, JS isn't actually that awesome for "shell" scripting, so I am
relying on a couple of third-party modules -- rimraf for an equivalent
to `rm -rf` and node-find-files to stand in for `find` -- which I chose
based on the fact that we've previously used them at FB (at least,
copies exist under `scripts/third_party/node_modules`).