static_upstream fixups
Summary: See Title.

The most pervasive new type of error is one where object property slots that started out with a `null` initialization must now specify what that `null` represents the absence of. We kind of did this in a few places before (class properties, mostly) -- but now we're more consistent about it.

Reviewed By: mroch

Differential Revision: D2570720

fb-gh-sync-id: b4ef6bee861b7ee954ff46d64948ca37a206ff29