Developed distinct methods for generating storage, serialzation, range behavior, and debug keys from fields
Summary: Previously, we were using `RelayQueryField::getStorageKey()` to do a lot of things. Its output at once produced:

* Storage keys for `RelayRecordStore`
* Range behavior keys for use in mutations
* Invariant messages
* Serialization keys for fields

The demands placed on this method made it difficult to change it without fear of breaking other things. For instance, it's desirable to make sure that `RelayRecordStore`'s storage keys are stable, no matter the order and nature of the source field arguments. To change (eg. sort and stably stringify) the storage key, would affect the other three consumers. Worse, to change the behavior even further down at the `printRelayQueryCall` level would break the actual queries that get sent to the server.

This diff adds two new methods to `RelayQuery`:

* `getRangeBehaviorKey()` (prints just the arguments, using legacy call syntax, without the leading period, in a stable order)
* `getDebugNameWithCalls()` (prints out the field name and the args using legacy call syntax, separated by a period)

…and changes the behavior of `getStorageKey()` and `getSerializationKey()` such that they produce identical keys no matter the order of their call arguments.

Reviewed By: josephsavona

Differential Revision: D2543114

fb-gh-sync-id: fa506132eb019b2c2e4236380e77441d94a13b9c