Hooking Up Incremental Read
Summary: Adding `readRelayDiskCache`, which uses the recently added `findRelayQueryLeaves` to figure
which nodes we need to read out from disk cache. It's essentially a BFT of the queries nodes and reading nodes out from disk recursively as they are needed.
`readRelayDiskCache` will keep track of pending node reads and pending rootcalls.
The pending maps ensure we don't waste time trying to read the same node from cache multiple times.
I also added a bunch of `hasFailed` to ensure we short circuit as fast as possible if we realized we are missing
data to avoid doing extra work.

Reviewed By: josephsavona

Differential Revision: D2548754

fb-gh-sync-id: 75d30bf4331dcf217b8396357dcdf39257b8b8e9