Fix master tests in Relay OSS, really
I actually fixed this in two ways, first by adding a `setImmediate` shim
(which worked) and then by switching to `resolveImmediate` (which didn't
work, but I failed to adequately test this). It seems the tests are very
sensitive, with their use of `jest.runAll{,Pending}Timers`.

Third time lucky; let's just use `setTimeout`, which should be fine as
this is just the test suite.