Fix master tests in Relay OSS, really
Summary: I actually fixed this in two ways, first by adding a `setImmediate` shim (which worked) and then by switching to `resolveImmediate` (which didn't work, but I failed to adequately test this). It seems the tests are very sensitive, with their use of `jest.run{All,OnlyPending}Timers`.

Third time lucky; let's just use `setTimeout`, which should be fine as this is just the test suite.
Closes https://github.com/facebook/relay/pull/513

Reviewed By: josephsavona

Differential Revision: D2579979

fb-gh-sync-id: 888021c83e2cef41ac27e44597888b45d88ba91f