Compile Relay.Query tags to RelayQuery nodes
Summary: The `Relay.QL` and `RelayQL` tags convert to `GraphQL` nodes (which do not know the value of variables), but all `RelayStore` APIs use `RelayQuery` objects that are a combination of GraphQL nodes plus variables.

This adds the `Relay.Query` tag to create RelayQuery objects with empty variables (all argument values must be inline in the query).

Closes https://github.com/facebook/relay/pull/519

Reviewed By: yungsters

Differential Revision: D2585606

fb-gh-sync-id: 3e06c1d47d8aff4efb064eddbb7549c4cd0aab34