Relay: GraphQL OSS Syntax for Optimistic Payloads
Summary: Refactors `inferRelayFieldsFromData` (which is used to generate a query given an optimistic payload) to use GraphQL OSS syntax for fields with arguments instead of GraphQL FB with calls. The old syntax wil continue working, but it will now warn.

We do not have a full fledged GraphQL parser in Relay, yet. This works around it by using regex and `JSON.parse` magic to fake support for parsing GraphQL arguments.

Reviewed By: josephsavona

Differential Revision: D2589522

fb-gh-sync-id: c9c661ff5d1eb177fcd0133a737899dfb9675946