serialize GraphQL nodes as plain objects
Summary: This is a precursor to replacing the `GraphQL.*` representation of queries with plain objects. In this diff, the two methods of serializing query objects are consolidated into one and this serialization format is changed to match the intended plain object representation for queries.

Overview of  changes:
- Remove `RelayQuerySerializer` which is only used by `RelayQueryPath`
- Change `RelayQueryPath` to use `toGraphQL`, and rely on that classes' `toJSON` implementations
- Change `GraphQL.*` classes `toJSON` implementations to output objects in the planned plain object query representation, and change the `fromJSON` methods to be basically a no-op.

Note that the current GraphQL constructors only purpose is to transform the fields into a form usable by `RelayQuery`: the new toJSON methods output exactly the format that is expected, so that `fromJSON` requires no further transformation.

Reviewed By: yungsters

Differential Revision: D2608960

fb-gh-sync-id: 8dc23098c15269b7dedd25507cbe411450c8a973