Fix issue where queries with multiple fields on the same Connection get sorted out of order
Summary: This is related to #478, where having multiple fields that query for different ranges on a Connection come back in random order causing the RelayQueryWriter to get confused and fail to build all the edges.

It is also related to #490. The difference between the PR and #490 is that this request uses a type first sorting that does not otherwise change the order of the elements in the array.
Closes https://github.com/facebook/relay/pull/523

Reviewed By: yungsters

Differential Revision: D2600306

fb-gh-sync-id: e9286fdc0b423150d77b846030f6f07714ca0d27