Exclude nodes that should be skipped because of directives in getChildren
Summary: The wording of the spec regarding `skip` and `include` was ambiguous so I clarified with leebyron and dschafer and updated it to reflect that a field should be skipped if *either* `skip: true || include: false` - see http://facebook.github.io/graphql/#sec--include.

Original PR Summary:

Fixes #536

I've been working with graphql-ruby and relay and been wanting to use directives. I found #536 yesterday night and tried to work on a solution today.

Following what josephsavona said in the issue, I've modified`RelayQueryNode.getChildren` so it only return children that are not skipped (meaning `skip if: true` and `include if: false`). I've implemented the function `RelayQueryNode.shouldBeIncluded` which
  - Checks if a skip directive is present ( It has priority over include)
     - If an if argument is present, return the negated value
  - Checks if an include directive is present
     - if an if argument, return the value
  - return true in all other cases

I've tried to take into consideration that other directives and arguments other than `if:` might be added, which might make the code a little bit more complicated, let me know if I should just assume that the argument is if right now.

Let me know if the solution makes sense. I've added 2 tests, modified a printing test that
Closes https://github.com/facebook/relay/pull/545

Reviewed By: yungsters

Differential Revision: D2612773

fb-gh-sync-id: 439ee42512abcfd41b373afa81b3001d0bd99ac7