Relay: Remove Value Substitution Warning
Summary: There are valid applications of value (non-variable) substitutions such as using constants. The pitfall to beware of here is that for container fragments, the values will be memoized the first time the fragment is evaluated. As such, the value of the varaible should not change (or else the behavior of the fragment may be misleading).

Reviewed By: josephsavona

Differential Revision: D2648096

fb-gh-sync-id: c5fe0a79baba8baa1922f08a98d74ce00922a71f