Relax buildRQL invariant for empty queries
Summary: Relay detects the use of abbreviated query definitions (##() => Relay.QL`query { node(id: $id) }`##) via the function argument length. Some environments wrap the function, so this reverts to the behavior pre-D2618129 and only uses the abbreviated form handler in the case that the query does not already contain a fragment.

Reviewed By: yungsters

Differential Revision: D2657418

fb-gh-sync-id: 1bf28e6ec5d9a18fd824a39a72138cbe922688fd