Add `Relay.Store.applyUpdate` to create a transaction w/o comitting it
Summary: For #550

Adds a `Relay.Store.applyUpdate(mutation, callbacks)` similar to `update`, but only applies the optimistic update. It returns a transaction object that can be used to `commit()` the mutation to the server or `rollback()` the optimistic change.

There weren't any tests for update so I added a test for `applyUpdate` and `update`, just making sure if the correct functions on the Queue and Transaction are called.

Closes https://github.com/facebook/relay/pull/599

Reviewed By: yungsters

Differential Revision: D2661433

fb-gh-sync-id: 54d40b38f035128f6532e8ee7fa23676d3e7ef6c