Make RelayRenderer run queries after mount
...to make sure `RelayRenderer` does not run queries during synchronous server-side rendering.

In the `RelayRenderer` tests, I had to replace `ShallowRenderer.render` with the traditional `ReactDOM.render`, because `ShallowRenderer.render` does not invoke some component lifecycle methods including `componentDidMount` that now run queries. But the logic of the tests was not changed, as visible behaviour of `RelayRenderer` was fully retained.

Added a test to check that queries are not run on the server.