Track deferred fragment state by tracking whether a query is resolved rather than whether it's in-flight.
Summary: Presently, the deferred query tracker sets a flag when a query for a deferred fragment is in-flight, and clears that flag when it resolves. This diff inverts that architecture, resuting in a simpler one where a deferred fragment's ‘ready’ status (as told by `hasFragmentData()`) is stored on the record itself, at write-time.

Reviewed By: josephsavona

Differential Revision: D2647722

fb-gh-sync-id: 01bc1465eb6d06e6d253f9d0c244235213f889c5