Improve performance of writing Relay payloads
Summary: `RelayQueryWriter` currently traverses *all* fragments when processing a payload and attempts to write their fields, regardless of whether the result type and fragment type actually match. For fields with an abstract/union type, we should instead only follow fragments whose type matches the actual result type - this can have a significant effect on performance, especially in cases where the top-level query has a union with a high branching factor.

For example, the previous writer would traverse both fragments A & B in the below example. With these changes, only fragment A is traversed (the type of the record).

```
// query
fragment on Type {
  id
  __typename
  ... on A { ... }
  ... on B { ... }
}

// response
{
  id: "...",
  __typename: "A"
  ...
}
```

This necessitated the following other changes:
- Changes `RelayQueryPath` to record fragments along the path, so that they can be correctly followed/skipped on refetch.
- Changes `RelayStoreData.buildFragmentQueryForDataID` to no longer refragment queries created for `setVariables`. This preserves fragments better and is also faster (avoids another full tree traversal).
- Adds `isConcrete` metadata to fragments of a concrete type; abstract fragments must always be traversed (skipping them would require understanding the full inheritance tree of the schema which isn't worth it). The default for runtime generated fragments is to assume they are an abstract type, which ensures that only known-safe fragments will be skipped.

Closes https://github.com/facebook/relay/pull/571

Reviewed By: yungsters

Differential Revision: D2631911

fb-gh-sync-id: 1ee19e2b8138290731cb4e8582c14f1abb53b919